requirements.txt
Django==3.1.7
Pillow==8.2.0
django-allauth==0.44.0
django-environ==0.4.5
whitenoise==5.2.0
gunicorn==20.0.4
psycopg2==2.8.5


